[MODIFY] 로그인 페이지 헤더, 타이틀 컴포넌트화
