+++ src/components/LoginComponents/LoginHeader.jsx
@@ -1,0 +1,11 @@
+import { Icon } from '../Icon/index.js';
+
+import * as S from './LoginHeader.style.jsx';
+
+export default function LoginHeader() {
+    return (
+        <S.Header>
+            <Icon id='logo' width='283.65px' height='39.93px' />
+        </S.Header>   
+    );
+}
